Clean up tests
Tests for public APIs should not be embedded in source files for
several reasons:

1. It introduces the possibility that private APIs are accidentally
   used by the tests in ways which are impossible for an actual
   consumer of the public API.
2. It is not conventional, so redundant tests end up being created
   (see redundant tests that showed up for the `Read::read_exact`
   and `Read::bytes` APIs) because attention is not paid to the
   fact that there are already tests in the tests.
3. It is more difficult to exclude tests from code coverage
   analysis, which means (once there is branch coverage) the
   overall code coverage level is incorrect due to including test
   files in the calculation.
