+++ binrw/src/endian.rs
@@ -34,20 +34,3 @@ impl Default for Endian {
         Endian::Native
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn endian_to_string() {
-        for &(ref endian, string) in [
-            (Endian::Big, "Big"),
-            (Endian::Little, "Little"),
-            (Endian::Native, "Native"),
-        ].iter() {
-            let converted: String = endian.into();
-            assert_eq!(converted, string)
-        }
-    }
-}
