+++ binrw/src/io/no_std/error.rs
@@ -1,8 +1,8 @@
+use core::fmt;
+
 pub struct Error {
     repr: Repr
 }
-
-use core::fmt;
 
 impl fmt::Debug for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -49,5 +49,11 @@ impl Error {
         match self.repr {
             Repr::Simple(kind) => kind,
         }
+    }
+}
+
+impl From<ErrorKind> for Error {
+    fn from(kind: ErrorKind) -> Self {
+        Self { repr: Repr::Simple(kind) }
     }
 }
