+++ binrw/tests/derive/enum.rs
