+++ binrw/tests/derive/mod.rs
@@ -1,0 +1,6 @@
+mod r#enum;
+mod struct_generic;
+mod struct_map;
+mod r#struct;
+mod unit_enum;
+mod unit_struct;
