+++ binrw/tests/derive/struct.rs
