--- binrw/tests/derive_tests.rs
@@ -1,8 +1,0 @@
-mod derive {
-    mod enums;
-    mod structs;
-    mod struct_generic;
-    mod struct_map;
-    mod unit_enum;
-    mod unit_struct;
-}
