+++ binrw/tests/io/mod.rs
@@ -1,1 +1,2 @@
+#[cfg(not(feature = "std"))]
 mod no_std;
