+++ binrw/tests/lib.rs
@@ -1,1 +1,2 @@
+mod derive;
 mod io;
