+++ binrw_derive/src/lib.rs
@@ -5,51 +5,21 @@ mod codegen;
 mod parser;
 
 use codegen::generate_impl;
-use parser::{Input, is_binread_attr};
+use parser::{Input, ParseResult, is_binread_attr};
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::{DeriveInput, parse_macro_input};
 
 #[proc_macro_derive(BinRead, attributes(binread, br))]
+#[cfg(not(tarpaulin_include))]
 pub fn derive_binread_trait(input: TokenStream) -> TokenStream {
-    derive_binread_internal(parse_macro_input!(input as DeriveInput)).into()
-}
-
-fn derive_binread_internal(input: DeriveInput) -> proc_macro2::TokenStream {
-    let binread_input = Input::from_input(&input);
-    generate_impl(&input, &binread_input)
+    derive_from_input(&parse_macro_input!(input as DeriveInput)).1.into()
 }
 
 #[proc_macro_attribute]
+#[cfg(not(tarpaulin_include))]
 pub fn derive_binread(_: TokenStream, input: TokenStream) -> TokenStream {
-    let mut derive_input = parse_macro_input!(input as DeriveInput);
-    let binread_input = Input::from_input(&derive_input);
-    let generated_impl = generate_impl(&derive_input, &binread_input);
-    let binread_input = binread_input.ok();
-
-    clean_struct_attrs(&mut derive_input.attrs);
-
-    match &mut derive_input.data {
-        syn::Data::Struct(input_struct) => {
-            clean_field_attrs(&binread_input, 0, &mut input_struct.fields);
-        },
-        syn::Data::Enum(input_enum) => {
-            for (index, variant) in input_enum.variants.iter_mut().enumerate() {
-                clean_struct_attrs(&mut variant.attrs);
-                clean_field_attrs(&binread_input, index, &mut variant.fields);
-            }
-        },
-        syn::Data::Union(union) => {
-            for field in union.fields.named.iter_mut() {
-                clean_struct_attrs(&mut field.attrs);
-            }
-        },
-    }
-
-    quote!(
-        #derive_input
-        #generated_impl
-    ).into()
+    derive_from_attribute(parse_macro_input!(input as DeriveInput)).into()
 }
 
 fn clean_field_attrs(binread_input: &Option<Input>, variant_index: usize, fields: &mut syn::Fields) {
@@ -80,30 +50,67 @@ fn clean_struct_attrs(attrs: &mut Vec<syn::Attribute>) {
     attrs.retain(|attr| !is_binread_attr(attr));
 }
 
+fn derive_from_attribute(mut derive_input: DeriveInput) -> proc_macro2::TokenStream {
+    let (binread_input, generated_impl) = derive_from_input(&derive_input);
+    let binread_input = binread_input.ok();
+
+    clean_struct_attrs(&mut derive_input.attrs);
+
+    match &mut derive_input.data {
+        syn::Data::Struct(input_struct) => {
+            clean_field_attrs(&binread_input, 0, &mut input_struct.fields);
+        },
+        syn::Data::Enum(input_enum) => {
+            for (index, variant) in input_enum.variants.iter_mut().enumerate() {
+                clean_struct_attrs(&mut variant.attrs);
+                clean_field_attrs(&binread_input, index, &mut variant.fields);
+            }
+        },
+        syn::Data::Union(union) => {
+            for field in union.fields.named.iter_mut() {
+                clean_struct_attrs(&mut field.attrs);
+            }
+        },
+    }
+
+    quote!(
+        #derive_input
+        #generated_impl
+    )
+}
+
+fn derive_from_input(derive_input: &DeriveInput) -> (ParseResult<Input>, proc_macro2::TokenStream) {
+    let binread_input = Input::from_input(&derive_input);
+    let generated_impl = generate_impl(&derive_input, &binread_input);
+    (binread_input, generated_impl)
+}
+
 #[cfg(test)]
-mod tests {
+#[cfg(tarpaulin)]
+#[test]
+fn derive_code_coverage_for_tarpaulin() {
     use runtime_macros_derive::emulate_derive_expansion_fallible;
     use std::{env, fs};
 
-    #[test]
-    fn derive_code_coverage() {
-        let derive_tests_folder = env::current_dir().unwrap()
-                .join("..")
-                .join("binrw")
-                .join("tests")
-                .join("derive");
+    let derive_tests_folder = env::current_dir().unwrap()
+        .join("..")
+        .join("binrw")
+        .join("tests")
+        .join("derive");
 
-        let mut run_success = true;
-        for entry in fs::read_dir(derive_tests_folder).unwrap() {
-            let entry = entry.unwrap();
-            if !entry.file_type().unwrap().is_file() {
-                continue
-            }
+    let mut run_success = true;
+    for entry in fs::read_dir(derive_tests_folder).unwrap() {
+        let entry = entry.unwrap();
+        if entry.file_type().unwrap().is_file() {
             let file = fs::File::open(entry.path()).unwrap();
-            let is_ok = emulate_derive_expansion_fallible(file, "BinRead", super::derive_binread_internal).is_ok();
-            run_success &= is_ok;
+            if emulate_derive_expansion_fallible(
+                file, "BinRead",
+                |input| derive_from_input(&input).1
+            ).is_err() {
+                run_success = false;
+            }
         }
-
-        assert!(run_success)
     }
+
+    assert!(run_success)
 }
